[fix] Minimal semi-working example

+++ core/Runtime/RtObjects/RtMember.h
@@ -8,5 +8,5 @@ enum class MemberType : uint32_t;
 struct RtMember {
     uint32_t name_hash;
     MemberType type;
-    uint64_t offset;
+    uint32_t offset;
 };
